R6 class representing a simulation experiment.

A simulation experiment with any number of DGPs, Methods, Evaluators, and Plotters.

Details

When run, an Experiment seamlessly combines DGPs and Methods, computing results in parallel. Those results can then be evaluated using Evaluators and plotted using Plotters.

Public fields

n_reps

The number of replicates of the Experiment.

name

The name of the Experiment.

Methods

Public methods

Experiment$new()

Experiment$run()

Experiment$generate_data()

Experiment$fit()

Experiment$evaluate()

Experiment$plot()

Experiment$create_doc_template()

Experiment$create_rmd()

Experiment$has_parent()

Experiment$get_parent()

Experiment$set_parent()

Experiment$add_child()

Experiment$get_children()

Experiment$get_descendants()

Experiment$add_dgp()

Experiment$update_dgp()

Experiment$get_dgps()

Experiment$add_method()

Experiment$update_method()

Experiment$get_methods()

Experiment$add_evaluator()

Experiment$update_evaluator()

Experiment$get_evaluators()

Experiment$add_plot()

Experiment$update_plot()

Experiment$get_plots()

Experiment$add_vary_across()

Experiment$update_vary_across()

Experiment$remove_vary_across()

Experiment$get_vary_across()

Experiment$get_save_dir()

Experiment$set_save_dir()

Experiment$clone()

Method new()

Create a new Experiment.

Usage

Experiment$new(
  n_reps,
  name = "experiment",
  dgp_list = list(),
  method_list = list(),
  evaluator_list = list(),
  plotter_list = list(),
  parent = NULL,
  save_dir = NULL,
  ...
)

Arguments

n_reps

The number of replicates of the Experiment.

name

The name of the Experiment.

dgp_list

An optional list of DGP objects.

method_list

An optional list of Method objects.

evaluator_list

An optional list of Evaluator objects.

plotter_list

An option list of Plotter objects.

parent

An optional parent Experiment object to extend.

save_dir

An optional directory in which to save the experiment's results.

...

Not used.

Returns

A new Experiment object.

Method run()

Run the entire simulation experiment pipeline.

Usage

Experiment$run(
  parallel_strategy = c("reps", "dgps", "methods"),
  trial_run = FALSE,
  use_cached = FALSE,
  save = FALSE,
  ...
)

Arguments

parallel_strategy

A vector with some combination of "reps", "dgps", or "methods". Determines how computation will be distributed across available resources.

trial_run

If TRUE, run 1 rep of the simulation experiment.

use_cached

If TRUE, find and return previously saved results.

save

If TRUE, save results to disk.

...

Not used.

Returns

A list of results from the simulation experiment.

Method generate_data()

Generate data from each of the DGPs in the Experiment.

Usage

Experiment$generate_data(n_reps = 1)

Arguments

n_reps

The number of datasets to generate per DGP.

Returns

A list of length equal to the number of DGPs in the Experiment. Each element is a list of datasets of length n_reps.

Method fit()

Return fit results from the Methods in the Experiment.

Usage

Experiment$fit(
  parallel_strategy = c("reps", "dgps", "methods"),
  trial_run = FALSE,
  use_cached = FALSE,
  save = FALSE,
  ...
)

Arguments

parallel_strategy

A vector with some combination of "reps", "dgps", or "methods". Determines how computation will be distributed across available resources.

trial_run

If TRUE, run 1 rep of the simulation experiment.

use_cached

If TRUE, find and return previously saved results.

save

If TRUE, save results to disk.

...

Not used.

Returns

A list of results from the simulation experiment.

Method evaluate()

Evaluate the Experiment.

Usage

Experiment$evaluate(fit_results, use_cached = FALSE, save = FALSE, ...)

Arguments

fit_results

A list of results, as returned by the run method.

use_cached

If TRUE, find and return previously saved evaluation results.

save

If TRUE, save evaluation results to disk.

...

Not used.

Returns

A list of evaluation results.

Method plot()

Plot the simulation experiment's method results or evaluation results.

Usage

Experiment$plot(
  fit_results = NULL,
  eval_results = NULL,
  use_cached = FALSE,
  save = FALSE,
  ...
)

Arguments

fit_results

A list of results, as returned by the fit method.

eval_results

A list of results, as returned by the evaluate method.

use_cached

If TRUE, find and return previously saved results.

save

If TRUE, save plots to disk.

...

Not used.

Returns

A list of plots.

Method create_doc_template()

Usage

Experiment$create_doc_template(...)

Method create_rmd()

Usage

Experiment$create_rmd(open = TRUE, ...)

Method has_parent()

Usage

Experiment$has_parent()

Method get_parent()

Usage

Experiment$get_parent()

Method set_parent()

Usage

Experiment$set_parent(parent)

Method add_child()

Usage

Experiment$add_child(child, name = NULL, ...)

Method get_children()

Usage

Experiment$get_children()

Method get_descendants()

Usage

Experiment$get_descendants(include_self = FALSE)

Method add_dgp()

Usage

Experiment$add_dgp(dgp, name = NULL, ...)

Method update_dgp()

Usage

Experiment$update_dgp(dgp, name, ...)

Method get_dgps()

Usage

Experiment$get_dgps()

Method add_method()

Usage

Experiment$add_method(method, name = NULL, ...)

Method update_method()

Usage

Experiment$update_method(method, name, ...)

Method get_methods()

Usage

Experiment$get_methods()

Method add_evaluator()

Usage

Experiment$add_evaluator(evaluator, name = NULL, ...)

Method update_evaluator()

Usage

Experiment$update_evaluator(evaluator, name, ...)

Method get_evaluators()

Usage

Experiment$get_evaluators()

Method add_plot()

Usage

Experiment$add_plot(plotter, name = NULL, ...)

Method update_plot()

Usage

Experiment$update_plot(plotter, name, ...)

Method get_plots()

Usage

Experiment$get_plots()

Method add_vary_across()

Usage

Experiment$add_vary_across(dgp = NULL, method = NULL, param_name, param_values)

Method update_vary_across()

Usage

Experiment$update_vary_across(
  dgp = NULL,
  method = NULL,
  param_name,
  param_values
)

Method remove_vary_across()

Usage

Experiment$remove_vary_across()

Method get_vary_across()

Usage

Experiment$get_vary_across()

Method get_save_dir()

Usage

Experiment$get_save_dir()

Method set_save_dir()

Usage

Experiment$set_save_dir(save_dir)

Method clone()

The objects of this class are cloneable with this method.

Usage

Experiment$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.