R6 class representing a method.

A method to be fit and assessed in the Experiment.

Public fields

name

The name of the Method.

method_fun

The method function.

method_params

(Named) list of parameters to input into the method function.

Methods

Public methods

Method$new()

Method$fit()

Method$clone()

Method new()

Create a new Method.

Usage

Method$new(method_fun, name = NULL, ...)

Arguments

method_fun

The method function.

name

(Optional) The name of the Method.

...

Arguments to pass into method_fun().

Returns

A new Method object.

Method fit()

Fit a Method on data using the provided Method parameters.

Usage

Method$fit(data_list, ...)

Arguments

data_list

List of data to pass into method_fun(). If named, should match arguments in method_fun().

...

Arguments to pass into method_fun() that will overwrite the initialized Method parameters. If no additional arguments are provided, the Method will be fit using method_fun() and the parameters that were set when Method$new() was called.

Returns

Result of method_fun(), coerced into a single tibble row.

Method clone()

The objects of this class are cloneable with this method.

Usage

Method$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.