R6 class representing a plotter.

A plotter (or plotting function) to plot the performance of methods and/or its evaluation metrics from the Experiment run.

Public fields

plot_fun

The plotting function.

plot_params

(Named) list of parameters to input into the plotting function.

rmd_options

List of options to control the aesthetics of the Plotter's displayed plot in the knitted R Markdown report. Currently, possible options are "height" and "width" (in inches).

show

If TRUE (default), show the resulting plot in the R Markdown report; if FALSE, hide output in the R Markdown report.

Methods

Public methods

Plotter$new()

Plotter$plot()

Plotter$clone()

Method new()

Create a new Plotter.

Usage

Plotter$new(plot_fun, rmd_options = list(), ...)

Arguments

plot_fun

The plotting function.

rmd_options

List of options to control the aesthetics of the Plotter's displayed plot in the knitted R Markdown report. Currently, possible options are "height" and "width" (in inches).

...

Arguments to pass into plot_fun().

Details

When plotting or running the Experiment (see Experiment$plot() or Experiment$run()), the named arguments fit_results, eval_results, and vary_param are automatically passed into the Plotter function plot_fun() and serve as placeholders for the Experiment$fit() results, the Experiment$evaluate() results, and the name of the varying parameter, respectively. To plot the performance of a method(s) fit and/or its evaluation metrics then, the Plotter function plot_fun() should take in the named arguments fit_results and/or eval_results. See Experiment$fit() or fit_experiment() for details on the format of fit_results. See Experiment$evaluate() or evaluate_experiment() for details on the format of eval_results. If the Plotter is used for Experiments with varying parameters, vary_param should be used as a stand in for the name of this varying parameter.

Returns

A new Plotter object.

Method plot()

Plot the performance of methods and/or their evaluation metrics from the Experiment using the Plotter and the provided parameters.

Usage

Plotter$plot(fit_results = NULL, eval_results = NULL, vary_param = NULL, ...)

Arguments

fit_results

A tibble, typically returned by the Experiment$fit() method.

eval_results

A list of tibbles, typically returned by the Experiment$evaluate() method.

vary_param

Name of parameter/argument that was varied in the Experiment, i.e., Experiment$get_vary_across()$param_name. Use NULL (default) if no vary_across component in Experiment run.

...

Not used.

Returns

Result of plot_fun(), which should be a plot.

Method clone()

The objects of this class are cloneable with this method.

Usage

Plotter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.