R6 class representing a simulation experiment.

A simulation experiment with any number of DGPs, Methods, Evaluators, and Plotters.

Details

When run, an Experiment seamlessly combines DGPs and Methods, computing results in parallel. Those results can then be evaluated using Evaluators and plotted using Plotters.

Public fields

name

The name of the Experiment.

Methods

Public methods

Experiment$new()

Experiment$run()

Experiment$generate_data()

Experiment$fit()

Experiment$evaluate()

Experiment$plot()

Experiment$create_doc_template()

Experiment$create_rmd()

Experiment$add_dgp()

Experiment$update_dgp()

Experiment$remove_dgp()

Experiment$get_dgps()

Experiment$add_method()

Experiment$update_method()

Experiment$remove_method()

Experiment$get_methods()

Experiment$add_evaluator()

Experiment$update_evaluator()

Experiment$remove_evaluator()

Experiment$get_evaluators()

Experiment$add_plotter()

Experiment$update_plotter()

Experiment$remove_plotter()

Experiment$get_plotters()

Experiment$add_vary_across()

Experiment$update_vary_across()

Experiment$remove_vary_across()

Experiment$get_vary_across()

Experiment$set_rmd_options()

Experiment$get_save_dir()

Experiment$set_save_dir()

Experiment$save()

Experiment$print()

Experiment$clone()

Method new()

Create a new Experiment.

Usage

Experiment$new(
  name = "experiment",
  dgp_list = list(),
  method_list = list(),
  evaluator_list = list(),
  plotter_list = list(),
  clone_from = NULL,
  save_dir = NULL,
  ...
)

Arguments

name

The name of the Experiment.

dgp_list

An optional list of DGP objects.

method_list

An optional list of Method objects.

evaluator_list

An optional list of Evaluator objects.

plotter_list

An optional list of Plotter objects.

clone_from

An optional Experiment object to use as a base for this one.

save_dir

An optional directory in which to save the experiment's results. If NULL, results are saved at ./results/{Experiment$name}

...

Not used.

Returns

A new Experiment object.

Method run()

Run the entire simulation experiment pipeline (fitting, evaluating, and plotting)

Usage

Experiment$run(
  n_reps = 1,
  parallel_strategy = c("reps", "dgps", "methods"),
  use_cached = FALSE,
  save = FALSE,
  verbose = 1,
  ...
)

Arguments

n_reps

The number of replicates of the Experiment for this run.

parallel_strategy

A vector with some combination of "reps", "dgps", or "methods". Determines how computation will be distributed across available resources.

use_cached

If TRUE, find and return previously saved results. If cached results cannot be found, run experiment anyways. Can also be a vector with some combination of "fit", "eval", or "plot" to use their respective cached results.

save

If TRUE, save results to disk. Can also be a vector with some combination of "fit", "eval", or "plot" to save to disk.

verbose

Level of verboseness. Default is 1, which prints out messages after major checkpoints in the experiment. If 0, no messages are printed.

...

Not used.

Returns

A list of results from the simulation experiment

fit_results

A tibble containing results from the fit method. In addition to results columns, has columns named 'rep', 'dgp', 'method', and the vary_across parameter name if applicable.

eval_results

A list of tibbles containing results from the evaluate method, which evaluates each Evaluator in the Experiment. Length of list is equivalent to the number of Evaluators.

plot_results

A list of tibbles containing results from the plot method, which plots each Plotter in the Experiment. Length of list is equivalent to the number of Plotters.

Method generate_data()

Generate data from each of the DGPs in the Experiment.

Usage

Experiment$generate_data(n_reps = 1, ...)

Arguments

n_reps

The number of datasets to generate per DGP.

...

Not used.

Returns

A list of length equal to the number of DGPs in the Experiment. Each element in the list is a list of n_reps datasets generated by the given DGP.

Method fit()

Fit all Methods in the Experiment and return results.

Usage

Experiment$fit(
  n_reps = 1,
  parallel_strategy = c("reps", "dgps", "methods"),
  use_cached = FALSE,
  save = FALSE,
  verbose = 1,
  ...
)

Arguments

n_reps

The number of replicates to run.

parallel_strategy

A vector with some combination of "reps", "dgps", or "methods". Determines how computation will be distributed across available resources.

use_cached

If TRUE, find and return previously saved results. If cached results cannot be found, fit experiment anyways.

save

If TRUE, save results to disk.

verbose

Level of verboseness. Default is 1, which prints out messages after major checkpoints in the experiment. If 0, no messages are printed.

...

Not used.

Returns

A tibble containing the results from fitting all Methods across all DGPs for n_reps repetitions. In addition to results columns, has columns named 'rep', 'dgp', 'method', and the vary_across parameter name if applicable.

Method evaluate()

Evaluate the performance of method fits across all Evaluators in the Experiment and return evaluation results.

Usage

Experiment$evaluate(
  fit_results,
  use_cached = FALSE,
  save = FALSE,
  verbose = 1,
  ...
)

Arguments

fit_results

A tibble, as returned by the fit method.

use_cached

If TRUE, find and return previously saved evaluation results. If cached results cannot be found, evaluate experiment anyways.

save

If TRUE, save evaluation results to disk.

verbose

Level of verboseness. Default is 1, which prints out messages after major checkpoints in the experiment. If 0, no messages are printed.

...

Not used.

Returns

A list of evaluation result tibbles, one for each Evaluator.

Method plot()

Plot the performance of methods and/or its evaluation metrics using all Plotters in the Experiment and return plot results.

Usage

Experiment$plot(
  fit_results = NULL,
  eval_results = NULL,
  use_cached = FALSE,
  save = FALSE,
  verbose = 1,
  ...
)

Arguments

fit_results

A tibble, as returned by the fit method.

eval_results

A list of result tibbles, as returned by the evaluate method.

use_cached

If TRUE, find and return previously saved results. If cached results cannot be found, plot experiment anyways.

save

If TRUE, save plots to disk.

verbose

Level of verboseness. Default is 1, which prints out messages after major checkpoints int the experiment. If 0, no messages are printed.

...

Not used.

Returns

A list of plots, one for each Plotter.

Method create_doc_template()

Create documentation template (a series of .md files) to fill out for the R Markdown results report. The documentation files can be found in the Experiment's results directory under docs/.

Usage

Experiment$create_doc_template(save_dir = self$get_save_dir(), ...)

Arguments

save_dir

An optional directory in which to find the experiment's saved results.

...

Not used.

Returns

The original Experiment object.

Method create_rmd()

Knits an R Markdown file summarizing the results of an Experiment. Outputs an R Markdown-generated html file, saved in the Experiment's root results directory.

Usage

Experiment$create_rmd(open = TRUE, verbose = 2, ...)

Arguments

open

If TRUE, open the R Markdown-generated html file in a web browser.

verbose

Level of verboseness (0, 1, 2) when knitting R Markdown. Default is 2.

...

Not used.

Returns

The original Experiment object.

Method add_dgp()

Add a new DGP to the Experiment.

Usage

Experiment$add_dgp(dgp, name = NULL, ...)

Arguments

dgp

A DGP object.

name

Name of DGP.

...

Not used.

Returns

The original Experiment object with the DGP added.

Method update_dgp()

Update an existing DGP in the Experiment.

Usage

Experiment$update_dgp(dgp, name, ...)

Arguments

dgp

A DGP object.

name

Name of existing DGP to update.

...

Not used.

Returns

The original Experiment object with the updated DGP.

Method remove_dgp()

Remove a DGP from the Experiment.

Usage

Experiment$remove_dgp(name, ...)

Arguments

name

Name of DGP to remove.

...

Not used.

Returns

The original Experiment object with the DGP removed.

Method get_dgps()

Get all DGPs in the Experiment.

Usage

Experiment$get_dgps()

Returns

A list of DGP objects in the Experiment.

Method add_method()

Add a new Method to the Experiment.

Usage

Experiment$add_method(method, name = NULL, ...)

Arguments

method

A Method object.

name

Name of Method.

...

Not used.

Returns

The original Experiment object with the Method added.

Method update_method()

Update an existing Method in the Experiment.

Usage

Experiment$update_method(method, name, ...)

Arguments

method

A Method object.

name

Name of existing Method to update.

...

Not used.

Returns

The original Experiment object with the updated Method.

Method remove_method()

Remove a Method from the Experiment.

Usage

Experiment$remove_method(name, ...)

Arguments

name

Name of Method to remove.

...

Not used.

Returns

The original Experiment object with the Method removed.

Method get_methods()

Get all Methods in the Experiment.

Usage

Experiment$get_methods()

Returns

A list of Method objects in the Experiment.

Method add_evaluator()

Add a new Evaluator to the Experiment.

Usage

Experiment$add_evaluator(evaluator, name = NULL, ...)

Arguments

evaluator

An Evaluator object.

name

Name of Evaluator.

...

Not used.

Returns

The original Experiment object with the Evaluator added.

Method update_evaluator()

Update an existing Evaluator in the Experiment.

Usage

Experiment$update_evaluator(evaluator, name, ...)

Arguments

evaluator

An Evaluator object.

name

Name of existing Evaluator to update.

...

Not used.

Returns

The original Experiment object with the updated Evaluator.

Method remove_evaluator()

Remove an Evaluator from the Experiment.

Usage

Experiment$remove_evaluator(name, ...)

Arguments

name

Name of Evaluator to remove.

...

Not used.

Returns

The original Experiment object with the Evaluator removed.

Method get_evaluators()

Get all Evaluators in the Experiment.

Usage

Experiment$get_evaluators()

Returns

A list of Evaluator objects in the Experiment.

Method add_plotter()

Add a new Plotter to the Experiment.

Usage

Experiment$add_plotter(plotter, name = NULL, ...)

Arguments

plotter

A Plotter object.

name

Name of Plotter.

...

Not used.

Returns

The original Experiment object with the Plotter added.

Method update_plotter()

Update an existing Plotter in the Experiment.

Usage

Experiment$update_plotter(plotter, name, ...)

Arguments

plotter

A Plotter object.

name

Name of existing Plotter to update.

...

Not used.

Returns

The original Experiment object with the updated Plotter.

Method remove_plotter()

Remove a Plotter from the Experiment.

Usage

Experiment$remove_plotter(name, ...)

Arguments

name

Name of Plotter to remove.

...

Not used.

Returns

The original Experiment object with the Plotter removed.

Method get_plotters()

Get all Plotters in the Experiment.

Usage

Experiment$get_plotters()

Returns

A list of Plotter objects in the Experiment.

Method add_vary_across()

Add a vary_across component to the Experiment. When the Experiment is run, the Experiment is systematically varied across values of the specified parameter in the DGP or Method while all other parameters are held constant at their baseline value.

Usage

Experiment$add_vary_across(dgp = NULL, method = NULL, param_name, param_values)

Arguments

dgp

Name of DGP to vary in the Experiment. Can also be a DGP object that matches one in the Experiment. Must provide either the dgp or method argument.

method

Name of Method to vary in the Experiment. Can also be a Method object that matches one in the Experiment. Must provide either the dgp or method argument.

param_name

Name of parameter to vary. Should match an argument in the DGP or Method function.

param_values

Atomic vector or list of values for the param_name to take on and vary across in the Experiment run.

Returns

The original Experiment object with the vary_across component added.

Method update_vary_across()

Update the vary_across component in the Experiment. Used if a vary_across component already exists in the Experiment.

Usage

Experiment$update_vary_across(
  dgp = NULL,
  method = NULL,
  param_name,
  param_values
)

Arguments

dgp

Name of DGP to vary in the Experiment. Can also be a DGP object that matches one in the Experiment. Must provide either the dgp or method argument.

method

Name of Method to vary in the Experiment. Can also be a Method object that matches one in the Experiment. Must provide either the dgp or method argument.

param_name

Name of parameter to vary. Should match an argument in the DGP or Method function.

param_values

Atomic vector or list of values for the param_name to take on and vary across in the Experiment run.

Returns

The original Experiment object with the updated vary_across component.

Method remove_vary_across()

Remove the vary_across component in the Experiment.

Usage

Experiment$remove_vary_across()

Returns

The original Experiment object with the vary_across component removed.

Method get_vary_across()

Get the vary_across component from the Experiment.

Usage

Experiment$get_vary_across()

Returns

A list with the following components

dgp, method

Name of the DGP/Method that was varied.

param_name

Name of parameter in the DGP/Method that was varied.

param_values

Atomic vector or list of values for the param_name to take on and vary across in the Experiment run.

Method set_rmd_options()

Set R Markdown options for displaying Evaluator or Plotter outputs in the summary report.

Usage

Experiment$set_rmd_options(field_name, name, show = NULL, ...)

Arguments

field_name

One of "evaluator" or "plotter".

name

Name of Evaluator or Plotter to set R Markdown options.

show

If TRUE, show output; if FALSE, hide output in R Markdown report. Default NULL does not change the "show" field in Evaluator/Plotter.

...

Named R Markdown options to set. If field_name = "plotter", options are "height" and "width". If field_name = "evaluator", see options for prettyDT().

Returns

The original Experiment object with the rmd_options and/or show fields modified in the Evaluator/Plotter.

Method get_save_dir()

Get the directory in which the Experiment's results are saved.

Usage

Experiment$get_save_dir()

Returns

The directory in which the Experiment's results are saved.

Method set_save_dir()

Set the directory in which the Experiment's results are saved.

Usage

Experiment$set_save_dir(save_dir)

Arguments

save_dir

The directory in which the Experiment's results will be saved.

Returns

The original Experiment object with the updated saving directory.

Method save()

Save an Experiment object.

Usage

Experiment$save()

Returns

The original Experiment object.

Method print()

Print an Experiment in a nice format, showing the names of all DGPs, Methods, Evaluators, and Plotters in addition to the directory where results are saved and the vary_across component (if any).

Usage

Experiment$print()

Returns

The original Experiment object.

Method clone()

The objects of this class are cloneable with this method.

Usage

Experiment$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.